Create user for integration with sumo logic
Currently we providing sumo logic with the heroku (admin) user to
allow it to read logs from our S3 buckets.  That's not a good idea
because heroku user has too much access.  This creates a specific
service user for sumo logic so that it will only have read access
to our S3 buckets because that's all it really needs.