BRIDGE-2020: Automate cloudtrail infrastructure
AWS Cloudtrail was manually setup on our bridge (prod) account. This
automates cloudtrail for both Bridge development and prod accounts.

This implementation is basically a copy of the cloudtrail example[1].
It creates a CloudTrail trail, an Amazon S3 bucket where logs are published,
and an Amazon SNS topic where notifications are sent. The bucket and topic
policies allow CloudTrail (from the specified regions) to publish logs to the
Amazon S3 bucket and to send notifications to an email that you specify.
Because CloudTrail automatically writes to the bucket_name/AWSLogs/account_ID/
folder, the bucket policy grants write privileges for that prefix.  We
also grant privileges to the Sumo service user so sumo can ingest
cloudtrail logs.

[1] https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-cloudtrail-trail.html