Move enforce mfa policy to bridge.yml
The enforce mfa policy lives in the bootstrap.yml which does not
get automatically deployed by travis. Moving it to the bridge.yml
file so that it gets updated by travis.

Note- this is the 1st change in a series to move this policy.
We need to duplcate the policy and then delete the old one
(in bootstrap.yml) in a follow on change.