Add CF template for default VPC
Bridge prod app is running in an AWS default VPC.  "Default" meaning that it
is the default VPC that comes with every AWS account.  Add a CF template
that can be used to re-create the "default" VPC if anything bad were to occur.
Also since bridge dev infra should be a mirror of prod we use this
template to create a VPC for bridge dev with a unique CIDR.  Our VPN
requires that every VPC have a unique CIDR.