[it-395] Fix aws config non-compliance (#111)
AWS config reports non-compliance in all our aws accounts due to
a misconfigured aws config rule.

In each aws account we have setup cloudtrail to send logs to a
central S3 bucket in our aws org-sagebase-logcentral account
(essentials-awss3cloudtrailbucket-1jz6pf8dzid7r). However the
aws config rule was setup to look for the cloudtrail logs in a
local bucket.

To fix the issue we re-configure the config rule to look for
logs in the central s3 bucket.