[IT-1053] Remove obsolete VPC and peering
bridge-develp vpc was created (a long time ago) as a mirror for the old bridge
prod vpc that only contained public subnets. It was created with the intention
of moving bridgedev env to the AWS bridge dev account. Since then @DwayneJengSage
moved bridge app into a new prod VPC with public & private subnets. The new
"BridgeServer2-develop-vpc" was created to mirror that new prod VPC. That means
that "bridge-develop" vpc is obsolete. To clear up the confusion I we delete
the obsolete "bridge-develop" vpc and its peering connection.